+++ src/controllers/request.controller.js
@@ -1,6 +1,6 @@
 import models from '../db/models';
 import sender from '../services/email.service';
-import Response from "../utils/response.utils";
+import Response from '../utils/response.utils';
 
 const { Request, User } = models;
 /**
@@ -48,12 +48,46 @@ export default class RequestController {
     // const requestData;
 
     await Request.findAll({ where: { user_id: req.body.user_id } })
-        .then((data) => {
-          if (data.length) {
-            return Response.Success(res, data, 200);
-          }
-          return Response.CustomError(res, 404, 'error', 'No requests found');
-        })
-        .catch((err) => Response.InternalServerError(res, err));
+      .then((data) => {
+        if (data.length) {
+          return Response.Success(res, data, 200);
+        }
+        return Response.CustomError(res, 404, 'error', 'No requests found');
+      })
+      .catch((err) => Response.InternalServerError(res, err));
+  }
+
+  /**
+     * @param {object} req
+     * @param {object} res
+     * @returns {json} request
+     */
+
+  static async createMultiCityRequest(req, res) {
+    try {
+      const { id: user_id } = req.currentUser.dataValues;
+      const {
+        category, origin, destination, departure_date, return_date, reason, room_id
+      } = req.body;
+      const bookingData = {
+        departure_date, return_date, user_id, room_id
+      };
+      const booking = await models.Booking.create(bookingData);
+      const { id: booking_id } = booking;
+      const requestData = {
+        user_id, category, origin, destination: destination.split(', '), departure_date, return_date, reason, booking_id
+      };
+      const request = await models.Request.create(requestData);
+      return res.status(201).json({
+        status: 'success',
+        data: { request, booking }
+      });
+    } catch (error) {
+      return res.status(500)
+        .json({
+          status: 'error',
+          error: 'Internal server error',
+        });
+    }
   }
 }
