+++ src/routes/api/index.router.js
@@ -1,8 +1,9 @@
 import express from 'express';
+
 const router = express.Router();
 
 router.get('/', (req, res) => {
-    res.status(200).json({ Message: 'Welcome! This is the NorthStar Barefoot Nomad homepage' });
+  res.status(200).json({ Message: 'Welcome! This is the NorthStar Barefoot Nomad homepage' });
 });
 
 export default router;
