--- src/routes/api/requests.js
@@ -1,13 +1,0 @@
-import express from 'express';
-import RequestController from '../../controllers/requests.controller';
-import authMiddleware from '../../middlewares/auth'
-
-const router = express.Router();
-
-const { createMultiCityRequest } = RequestController
-const { verifyUserToken } = authMiddleware
-/* Request routes */
-router.post('/multiCity', verifyUserToken, createMultiCityRequest )
-
-
-export default router;
