+++ src/routes/api/requests.router.js
@@ -17,4 +17,11 @@ router.patch('/request/:id',
 
 router.get('/requests', Auth.verifyToken, RequestController.findAll);
 
+/* Multi City Request route */
+
+const { createMultiCityRequest } = RequestController;
+const { verifyUserToken } = auth;
+
+router.post('/request/multiCity', verifyUserToken, createMultiCityRequest);
+
 export default router;
